sms_gateway/requirements.txt
Flask==3.0.3
google-cloud-pubsub==2.21.4
twilio==9.2.3
python-dotenv==1.0.1
gunicorn==22.0.0
